fix: 댓글 삭제 오류 해결

+++ TimeTableArtist/src/main/java/SamwaMoney/TimeTableArtist/Reply/service/ReplyService.java
@@ -62,11 +62,17 @@ public class ReplyService {
                 .orElseThrow(()->new EntityNotFoundException("해당 댓글이 존재하지 않습니다."));
     }
 
-    public void removeReply(Long replyId){
+    public String removeReply(Long memberId, Long replyId){
         Reply reply = findReplyById(replyId);
-        replyRepository.delete(reply);
-        Timetable timetable = findReplyById(replyId).getTimetable();
-        timetable.setReplyCount(timetable.getReplyCount() - 1);
+        if (Objects.equals(reply.getWriter().getMemberId(), memberId)){
+            Timetable timetable = findReplyById(replyId).getTimetable();
+            timetable.setReplyCount(timetable.getReplyCount()-1);
+            replyRepository.delete(reply);
+            return "댓글이 삭제되었습니다.";
+        }
+        else{
+            return "해당 댓글을 삭제할 권한이 없습니다.";
+        }
     }
 }
